V1.3

+++ backend/src/main/java/io/metersphere/performance/controller/PerformanceTestController.java
@@ -14,6 +14,7 @@ import io.metersphere.dto.DashboardTestDTO;
 import io.metersphere.dto.LoadTestDTO;
 import io.metersphere.dto.ScheduleDao;
 import io.metersphere.performance.service.PerformanceTestService;
+import io.metersphere.service.CheckOwnerService;
 import io.metersphere.service.FileService;
 import io.metersphere.track.request.testplan.*;
 import org.apache.shiro.authz.annotation.Logical;
@@ -35,6 +36,8 @@ public class PerformanceTestController {
     private PerformanceTestService performanceTestService;
     @Resource
     private FileService fileService;
+    @Resource
+    private CheckOwnerService checkOwnerService;
 
     @GetMapping("recent/{count}")
     public List<LoadTestDTO> recentTestPlans(@PathVariable int count) {
@@ -54,12 +57,14 @@ public class PerformanceTestController {
 
     @GetMapping("/list/{projectId}")
     public List<LoadTest> list(@PathVariable String projectId) {
+        checkOwnerService.checkProjectOwner(projectId);
         return performanceTestService.getLoadTestByProjectId(projectId);
     }
 
 
     @GetMapping("/state/get/{testId}")
     public LoadTest listByTestId(@PathVariable String testId) {
+        checkOwnerService.checkPerformanceTestOwner(testId);
         return performanceTestService.getLoadTestBytestId(testId);
     }
 
@@ -76,26 +81,31 @@ public class PerformanceTestController {
             @RequestPart("request") EditTestPlanRequest request,
             @RequestPart(value = "file", required = false) List<MultipartFile> files
     ) {
+        checkOwnerService.checkPerformanceTestOwner(request.getId());
         return performanceTestService.edit(request, files);
     }
 
     @GetMapping("/get/{testId}")
     public LoadTestDTO get(@PathVariable String testId) {
+        checkOwnerService.checkPerformanceTestOwner(testId);
         return performanceTestService.get(testId);
     }
 
     @GetMapping("/get-advanced-config/{testId}")
     public String getAdvancedConfiguration(@PathVariable String testId) {
+        checkOwnerService.checkPerformanceTestOwner(testId);
         return performanceTestService.getAdvancedConfiguration(testId);
     }
 
     @GetMapping("/get-load-config/{testId}")
     public String getLoadConfiguration(@PathVariable String testId) {
+        checkOwnerService.checkPerformanceTestOwner(testId);
         return performanceTestService.getLoadConfiguration(testId);
     }
 
     @PostMapping("/delete")
     public void delete(@RequestBody DeleteTestPlanRequest request) {
+        checkOwnerService.checkPerformanceTestOwner(request.getId());
         performanceTestService.delete(request);
     }
 
@@ -111,6 +121,7 @@ public class PerformanceTestController {
 
     @GetMapping("/file/metadata/{testId}")
     public List<FileMetadata> getFileMetadata(@PathVariable String testId) {
+        checkOwnerService.checkPerformanceTestOwner(testId);
         return fileService.getFileMetadataByTestId(testId);
     }
