+++ backend/src/main/java/io/metersphere/track/controller/TestPlanController.java
@@ -8,12 +8,14 @@ import io.metersphere.commons.constants.RoleConstants;
 import io.metersphere.commons.utils.PageUtils;
 import io.metersphere.commons.utils.Pager;
 import io.metersphere.commons.utils.SessionUtils;
+import io.metersphere.service.CheckOwnerService;
 import io.metersphere.track.dto.TestCaseReportMetricDTO;
 import io.metersphere.track.dto.TestPlanDTO;
 import io.metersphere.track.dto.TestPlanDTOWithMetric;
 import io.metersphere.track.request.testcase.PlanCaseRelevanceRequest;
 import io.metersphere.track.request.testcase.QueryTestPlanRequest;
 import io.metersphere.track.request.testplan.AddTestPlanRequest;
+import io.metersphere.track.request.testplancase.TestCaseRelevanceRequest;
 import io.metersphere.track.service.TestPlanProjectService;
 import io.metersphere.track.service.TestPlanService;
 import org.apache.shiro.authz.annotation.Logical;
@@ -31,6 +33,8 @@ public class TestPlanController {
     TestPlanService testPlanService;
     @Resource
     TestPlanProjectService testPlanProjectService;
+    @Resource
+    CheckOwnerService checkOwnerService;
 
     @PostMapping("/list/{goPage}/{pageSize}")
     public Pager<List<TestPlanDTO>> list(@PathVariable int goPage, @PathVariable int pageSize, @RequestBody QueryTestPlanRequest request) {
@@ -46,7 +50,7 @@ public class TestPlanController {
         QueryTestPlanRequest request = new QueryTestPlanRequest();
         request.setWorkspaceId(workspaceId);
         request.setProjectId(projectId);
-        return testPlanService.listTestPlan(request);
+        return testPlanService.listTestPlanByProject(request);
     }
 
     @PostMapping("/list/all")
@@ -69,6 +73,7 @@ public class TestPlanController {
 
     @PostMapping("/get/{testPlanId}")
     public TestPlan getTestPlan(@PathVariable String testPlanId) {
+        checkOwnerService.checkTestPlanOwner(testPlanId);
         return testPlanService.getTestPlan(testPlanId);
     }
 
@@ -80,19 +85,21 @@ public class TestPlanController {
 
     @PostMapping("/edit")
     @RequiresRoles(value = {RoleConstants.TEST_USER, RoleConstants.TEST_MANAGER}, logical = Logical.OR)
-    public void editTestPlan(@RequestBody TestPlan testPlan) {
-        testPlanService.editTestPlan(testPlan);
+    public void editTestPlan(@RequestBody TestPlanDTO testPlanDTO) {
+        testPlanService.editTestPlan(testPlanDTO);
     }
 
     @PostMapping("/edit/status/{planId}")
     @RequiresRoles(value = {RoleConstants.TEST_USER, RoleConstants.TEST_MANAGER}, logical = Logical.OR)
     public void editTestPlanStatus(@PathVariable String planId) {
+        checkOwnerService.checkTestPlanOwner(planId);
         testPlanService.editTestPlanStatus(planId);
     }
 
     @PostMapping("/delete/{testPlanId}")
     @RequiresRoles(value = {RoleConstants.TEST_USER, RoleConstants.TEST_MANAGER}, logical = Logical.OR)
     public int deleteTestPlan(@PathVariable String testPlanId) {
+        checkOwnerService.checkTestPlanOwner(testPlanId);
         return testPlanService.deleteTestPlan(testPlanId);
     }
 
@@ -108,11 +115,22 @@ public class TestPlanController {
 
     @GetMapping("/project/name/{planId}")
     public String getProjectNameByPlanId(@PathVariable String planId) {
+        checkOwnerService.checkTestPlanOwner(planId);
         return testPlanService.getProjectNameByPlanId(planId);
     }
 
-    @GetMapping("/project/{planId}")
-    public List<Project> getProjectByPlanId(@PathVariable String planId) {
-        return testPlanProjectService.getProjectByPlanId(planId);
+    @PostMapping("/project")
+    public List<Project> getProjectByPlanId(@RequestBody TestCaseRelevanceRequest request) {
+        List<String> projectIds = testPlanProjectService.getProjectIdsByPlanId(request.getPlanId());
+        request.setProjectIds(projectIds);
+        return testPlanProjectService.getProjectByPlanId(request);
+    }
+
+    @PostMapping("/project/{goPage}/{pageSize}")
+    public Pager<List<Project>> getProjectByPlanId(@PathVariable int goPage, @PathVariable int pageSize, @RequestBody TestCaseRelevanceRequest request) {
+        List<String> projectIds = testPlanProjectService.getProjectIdsByPlanId(request.getPlanId());
+        request.setProjectIds(projectIds);
+        Page<Object> page = PageHelper.startPage(goPage, pageSize, true);
+        return PageUtils.setPageInfo(page, testPlanProjectService.getProjectByPlanId(request));
     }
 }
